requirements.txt
aioredis==2.0.1
alembic==1.9.4
anyio==3.6.2
apache-airflow-client==2.5.1
asgiref==3.6.0
async-timeout==4.0.2
backoff==2.2.1
black==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cryptography==39.0.1
cymem==2.0.7
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
fastapi-jwt-auth==0.5.0
fastapi-pagination==0.11.4
googleapis-common-protos==1.56.2
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
Mako==1.2.4
MarkupSafe==2.1.2
msal==1.21.0
murmurhash==1.0.9
mypy-extensions==1.0.0
numpy==1.24.2
opentelemetry-api==1.16.0
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-exporter-jaeger-proto-grpc==1.16.0
opentelemetry-exporter-jaeger-thrift==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-asgi==0.37b0
opentelemetry-instrumentation-fastapi==0.37b0
opentelemetry-instrumentation-requests==0.37b0
opentelemetry-instrumentation-sqlalchemy==0.37b0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pathspec==0.11.0
pathy==0.10.1
platformdirs==3.0.0
preshed==3.0.8
protobuf==3.20.3
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
scikit-learn==1.2.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
SQLAlchemy==2.0.4
srsly==2.4.5
starlette==0.25.0
thinc==8.1.7
threadpoolctl==3.1.0
thrift==0.16.0
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
wasabi==1.1.1
wrapt==1.14.1
